Address review comments

+++ src/yb/util/tostring.h
@@ -188,7 +188,7 @@ std::string CollectionToString(const Collection& collection);
 template <class Collection, class Transform>
 std::string CollectionToString(const Collection& collection, const Transform& transform);
 
-std::string CStringArrayToString(char** elements, int length);
+std::string CStringArrayToString(char** elements, size_t length);
 
 template <class T>
 typename std::enable_if<yb_tostring::HasFreeFunction_to_string<T>::value,
