+++ src/yb/yql/pggate/ybc_pg_typedefs.h
@@ -476,9 +476,9 @@ typedef struct PgJwtAuthOptions {
   char* jwks;
   char* matching_claim_key;
   char** allowed_issuers;
-  int allowed_issuers_length;
+  size_t allowed_issuers_length;
   char** allowed_audiences;
-  int allowed_audiences_length;
+  size_t allowed_audiences_length;
   char* username;
   char* usermap;
 } YBCPgJwtAuthOptions;
